feat(ChromeDriver): implemented `getProperty()` method, to get the live value of an element property by using JavaScript
Selenium WebDriver API does note have a way to get the live value of an attribute, like an updated value of an input element. The only way to get the live value would be to use JavaScript. That is what the `getProperty()` method does. Integration tests written for this method, and passed as expected.