feat(Element,ChromeDriver): implemented rect, drag and drop on both element and driver methods
Integration test for `Element.dragAndDrop()` implemented and passed as expected. `dragAndDrop()` webdriver method is a dynamic method that can either use javascript when element is draggable to drag element and fire off javascript event listeners, or use `Element.dragAndDrop()` when element is not draggable to use webdriver actions API instead.